Read posts directly from BEEM and Grammar Checker.
Since using a web scrape to load the page took more time that needed the script now uses BEEM to grab the raw data directly from the blockchain. This adde the need to use regex to remove the HTML and Markdown syntax to get an accurate word count output.

Grammar checker uses language-checker via pip to read the stripped data for possoble grammar mistakes for the english language. Shoud add the ablity to pick th locale onec the user web app is created. This grammar checker is only an aproximation.